+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 ## 0.0.14 (2023-08-28)
 
 
+### Bug Fixes
+
+* build issue ([89b7bab](https://gitlab.com/paras205/boilerplate-react/commit/89b7bab7ed5f7fe63870961d67610dc0bfacf3c6))
+
+
 ### Features
 
 * add additional folder ([6a3e80e](https://gitlab.com/paras205/boilerplate-react/commit/6a3e80e0283bb533435461c064b30935cd953ddf))
-* added drawer ([48ffd43](https://gitlab.com/paras205/boilerplate-react/commit/48ffd43f4de23ef1bbc049550439bbec3817f69f))
-* added modal component ([9c03dd0](https://gitlab.com/paras205/boilerplate-react/commit/9c03dd0af58382034d9385190688637a8f34f6a6))
+* slider component ([150d29a](https://gitlab.com/paras205/boilerplate-react/commit/150d29abf913fd66d9e8fcc724e1c702a9114d31))